refactor: - currency symbols object moved to a constant in data folder - on currency change function created and passed to onChange select event handler - Styled Components moved to universal components folder
