+++ client/src/components/BoxWrapper/BoxWrapper.js
@@ -1,0 +1,12 @@
+import { Box as MuiBox, styled } from '@mui/material';
+
+const BoxWrapper = styled(MuiBox)(({ theme }) => ({
+  position: 'sticky',
+  top: 20,
+  padding: '20px 20px 30px',
+  borderRadius: 5,
+
+  border: `1px solid ${theme.palette.divider}`,
+}));
+
+export default BoxWrapper;
