+++ client/src/components/index.js
@@ -2,5 +2,10 @@ import AlertModal from './AlertModal/AlertModal';
 import SliderButton from './SliderButton/SliderButton';
 import LogOutModal from './LogOutModal/LogOutModal';
 import SnackBar from './SnackBar/SnackBar';
+import BoxWrapper from './BoxWrapper/BoxWrapper';
+import Section from './Section/Section';
+import Title from './Title/Title';
+import Select from './Select/Select';
+import MenuItem from './MenuItem/MenuItem';
 
-export { AlertModal, SliderButton, LogOutModal, SnackBar };
+export { AlertModal, SliderButton, LogOutModal, SnackBar, BoxWrapper, Section, Title, Select, MenuItem };
