+++ client/src/features/Cart/components/TotalInfoDialog/TotalInfoDialog.js
@@ -1,85 +1,26 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-
-import {
-  Button,
-  Stack,
-  Box as MuiBox,
-  Typography,
-  styled,
-  Select as MuiSelect,
-  MenuItem as MuiMenuItem,
-  alpha,
-} from '@mui/material';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-
 import { useNavigate } from 'react-router-dom';
 
+// MUI
+import { Button, Stack, Box as MuiBox, Typography } from '@mui/material';
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+import { BoxWrapper, Section, Title, Select, MenuItem } from '../../../../components';
+
+// Utils & Data
 import calcToursQuantityAndPrice from '../../utils/calcToursQuantityAndPrice';
 import currencyConverter from '../../../Tour/utils/currencyConverter';
-
-const currencySymbol = {
-  eur: '€',
-  usd: '$',
-  uah: '₴',
-};
-
-const BoxWrapper = styled(MuiBox)(({ theme }) => ({
-  position: 'sticky',
-  top: 90,
-  padding: '20px 20px 30px',
-  borderRadius: 5,
-
-  border: `1px solid ${theme.palette.divider}`,
-}));
-
-const Section = styled(MuiBox)(({ theme }) => ({
-  borderBottom: `1px solid ${theme.palette.divider}`,
-  paddingBottom: 15,
-  marginBottom: 20,
-}));
-
-const Title = styled((props) => <Typography variant="h3" color="text.primary" {...props} />)({
-  textTransform: 'uppercase',
-  fontWeight: 700,
-  marginBottom: 20,
-});
-
-const MenuItem = styled(MuiMenuItem)(({ theme }) => ({
-  transition: theme.transitions.create(['color', 'background-color']),
-
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.primary.main, 0.15),
-    color: theme.palette.primary.main,
-  },
-
-  '&.Mui-selected': {
-    color: theme.palette.primary.main,
-    fontWeight: 700,
-  },
-}));
-
-const Select = styled(MuiSelect)(({ theme }) => ({
-  marginBottom: '10px',
-  fontWeight: 700,
-  '.MuiOutlinedInput-notchedOutline': {
-    borderColor: `${theme.palette.divider}`,
-
-    transition: theme.transitions.create('border-color'),
-  },
-  '&:hover .MuiOutlinedInput-notchedOutline': {
-    borderColor: `${theme.palette.primary.main}`,
-  },
-  '.MuiSelect-icon': {
-    transition: theme.transitions.create('transform'),
-  },
-}));
+import currencySymbols from '../../../data/currencySymbols';
 
 const TotalInfoDialog = ({ cart }) => {
   const navigate = useNavigate();
   const [currency, setCurrency] = useState('eur');
 
   const { toursPrice, toursQuantity } = calcToursQuantityAndPrice(cart);
+
+  const onCurrencyChange = (e) => {
+    setCurrency(e.target.value);
+  };
 
   return (
     <BoxWrapper>
@@ -93,8 +34,8 @@ const TotalInfoDialog = ({ cart }) => {
 
       <Section>
         <Title>Currency</Title>
-        <Select fullWidth value={currency} onChange={(e) => setCurrency(e.target.value)} IconComponent={ExpandMoreIcon}>
-          {Object.keys(currencySymbol).map((key) => (
+        <Select fullWidth value={currency} onChange={onCurrencyChange} IconComponent={ExpandMoreIcon}>
+          {Object.keys(currencySymbols).map((key) => (
             <MenuItem key={key} value={key}>
               {key.toUpperCase()}
             </MenuItem>
@@ -106,7 +47,7 @@ const TotalInfoDialog = ({ cart }) => {
         <Stack direction="row" justifyContent="space-between">
           <Typography>To be paid:</Typography>
           <Typography fontSize={18}>
-            {currencySymbol[currency]}
+            {currencySymbols[currency]}
             {currencyConverter(toursPrice, currency)}
           </Typography>
         </Stack>
