+++ client/src/features/data/currencySymbols.js
@@ -1,0 +1,5 @@
+export default {
+  eur: '€',
+  usd: '$',
+  uah: '₴',
+};
