+++ client/src/store/slices/cartSlice/cartSlice.js
@@ -150,10 +150,6 @@ const cartSlice = createSlice({
     closeSnackBar: (state) => {
       state.snackBar.isSnackBarOpen = false;
     },
-
-    closeSnackBar: (state) => {
-      state.snackBar.isSnackBarOpen = false;
-    },
   },
 
   extraReducers: (builder) => {
